Refactored passport and authentication code
Moved sensitive data to dems.conf.json
Enforced indentation / trailing whites and commenting
Added controllers folder for semantics
