+++ .gitignore
@@ -1,3 +1,5 @@
 // Package Management
 node_modules
 public/libs/
+
+dems.conf.json
