--- config/keys.js
@@ -1,10 +1,0 @@
-// config/keys.js
-
-exports.googleKeys = 
-    {
-        clientID    : '155424153501-c8cops3ulln3irkq6lb77k2gcvgvuhlr.apps.googleusercontent.com',
-        clientSecret: 've_ALQ4Ex5wGiXHYwxSkpAh8',
-        callbackURL : 'http://localhost:3000/auth/google/callback',
-        realm       : 'http://localhost:3000/'
-    };
-
