+++ models/carer.js
@@ -1,9 +1,15 @@
-var Schema = require('jugglingdb').Schema;
-var schema = new Schema('mongodb');
+/**
+ * Carer Schema and Model
+ * Functionality for the Carer persistence.
+ */
 
-module.exports = schema.define('Carer', {
-    gid : { type: Number, index: true },
-    token : { type: String },
-    email : { type: String, limit: 150 },
-    name: { type: String, limit: 50 }
+var Schema = require('jugglingdb').Schema;
+
+var schemaMongo = new Schema('mongodb');
+
+module.exports = schemaMongo.define('Carer', {
+  gid : { type: Number, index: true },
+  token : { type: String },
+  email : { type: String, limit: 150 },
+  name: { type: String, limit: 50 }
 });
