+++ package.json
@@ -18,18 +18,18 @@
   },
   "homepage": "https://github.com/knotsoftware/dems-service",
   "dependencies": {
+    "body-parser": "~1.6.5",
     "bower": "~1.3.9",
+    "cookie-parser": "~1.3.2",
     "errorhandler": "~1.1.1",
     "express": "~4.8.3",
+    "express-session": "~1.7.6",
+    "jugglingdb": "~0.2.14",
+    "jugglingdb-mongodb": "0.0.4-4",
+    "method-override": "~2.1.3",
     "morgan": "~1.2.2",
-    "jugglingdb": ">= 0.0.0",
-    "jugglingdb-mongodb": ">= 0.0.0",
-    "passport": ">= 0.0.0",
-    "passport-google-oauth": ">= 0.0.0",
-    "express-session": ">= 0.0.0",
-    "cookie-parser": ">= 0.0.0",
-    "body-parser": ">= 0.0.0",
-    "method-override": ">= 0.0.0"
+    "passport": "~0.2.0",
+    "passport-google-oauth": "~0.1.5"
   },
   "devDependencies": {
     "grunt": "~0.4.5",
